net, driver, vritchroot: Separate tap specific and general helpers
In order to add additional devices, separate the tap and the other
general helpers, such that they can be reused.
