+++ pkg/network/driver/virtchroot/tap.go
@@ -22,13 +22,7 @@ package virtchroot
 import (
 	"os/exec"
 	"strconv"
-
-	"kubevirt.io/kubevirt/pkg/virt-handler/selinux"
 )
-
-type VirtCHRoot struct{}
-
-const virtChrootBin = "virt-chroot"
 
 func (v VirtCHRoot) AddTapDevice(name string, mtu int, queues int, ownerID int) error {
 	cmd := v.addTapDeviceCmd(name, mtu, queues, ownerID)
@@ -37,7 +31,7 @@ func (v VirtCHRoot) AddTapDevice(name string, mtu int, queues int, ownerID int) 
 
 func (v VirtCHRoot) AddTapDeviceWithSELinuxLabel(name string, mtu int, queues int, ownerID int, pid int) error {
 	cmd := v.addTapDeviceCmd(name, mtu, queues, ownerID)
-	return v.runWithSELinuxLabelFromPID(pid, cmd)
+	return runWithSELinuxLabelFromPID(pid, cmd)
 }
 
 func (v VirtCHRoot) addTapDeviceCmd(name string, mtu int, queues int, ownerID int) *exec.Cmd {
@@ -52,13 +46,4 @@ func (v VirtCHRoot) addTapDeviceCmd(name string, mtu int, queues int, ownerID in
 	}
 	// #nosec No risk for attacket injection. cmdArgs includes predefined strings
 	return exec.Command(virtChrootBin, cmdArgs...)
-}
-
-func (v VirtCHRoot) runWithSELinuxLabelFromPID(pid int, cmd *exec.Cmd) error {
-	ctxExec, err := selinux.NewContextExecutor(pid, cmd)
-	if err != nil {
-		return err
-	}
-
-	return ctxExec.Execute()
 }
