+++ pkg/network/driver/virtchroot/virtchroot.go
@@ -1,0 +1,39 @@
+/*
+ * This file is part of the KubeVirt project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * Copyright 2024 Red Hat, Inc.
+ *
+ */
+
+package virtchroot
+
+import (
+	"os/exec"
+
+	"kubevirt.io/kubevirt/pkg/virt-handler/selinux"
+)
+
+type VirtCHRoot struct{}
+
+const virtChrootBin = "virt-chroot"
+
+func runWithSELinuxLabelFromPID(pid int, cmd *exec.Cmd) error {
+	ctxExec, err := selinux.NewContextExecutor(pid, cmd)
+	if err != nil {
+		return err
+	}
+
+	return ctxExec.Execute()
+}
